megaloc/requirements.txt
# Python 3.10

pip>=23.2

# Torch
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Core scientific stack
numpy~=1.26
scipy>=1.11
Pillow>=10.0
opencv-python>=4.8
matplotlib==3.7.2

# Model backbone
timm==0.5.4   # upgrade later only if needed

# Jupyter (optional)
jupyterlab>=4.3
notebook>=7.0
ipywidgets>=8.1

# evo dependencies
argcomplete>=2.0
colorama>=0.4
